+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 - Travis realease on tag
 - Now plotting Flux callibration tables
 - Extra frequency axis for bandpass plot
+- Support for string field names in addition to field indices
 
 0.0.2
 -----